src/Backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
djangorestframework==3.11.0
drf-yasg==1.17.1
flex==6.14.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jsonpointer==2.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.4
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rfc3987==1.3.8
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
strict-rfc3339==0.7
uritemplate==3.0.1
urllib3==1.25.9
validate-email==1.3
django-email-utils==1.0.0
django-filter==2.2.0

gensim==3.8.3
numpy==1.18.4
pandas==1.0.4
collections
